# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yeshshree Operations API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/confirmations": {
      "post": {
        "tags": [
          "production"
        ],
        "summary": "Post Confirmation",
        "description": "200 (not 201): a retried client_ref returns the ORIGINAL result (invariant 4).",
        "operationId": "post_confirmation_api_v1_confirmations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "production"
        ],
        "summary": "List Confirmations",
        "description": "History incl. SAP outbox status (pending/batched/sent/\u2026 as sap_sync_status).",
        "operationId": "list_confirmations_api_v1_confirmations_get",
        "parameters": [
          {
            "name": "line_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Line Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfirmationRead"
                  },
                  "title": "Response List Confirmations Api V1 Confirmations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/confirmations/hold": {
      "post": {
        "tags": [
          "production"
        ],
        "summary": "Hold Confirmation",
        "description": "\u00a711.11 \u2014 no resolvable order: store the full body, nothing lost, nothing posted.",
        "operationId": "hold_confirmation_api_v1_confirmations_hold_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/confirmations/{confirmation_id}/correction": {
      "post": {
        "tags": [
          "production"
        ],
        "summary": "Request Correction",
        "operationId": "request_correction_api_v1_confirmations__confirmation_id__correction_post",
        "parameters": [
          {
            "name": "confirmation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Confirmation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__schemas__production__CorrectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__production__CorrectionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/confirmations/holds": {
      "get": {
        "tags": [
          "production"
        ],
        "summary": "List Holds",
        "operationId": "list_holds_api_v1_confirmations_holds_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "open",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldRead"
                  },
                  "title": "Response List Holds Api V1 Confirmations Holds Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/confirmations/holds/{hold_id}/resolve": {
      "post": {
        "tags": [
          "production"
        ],
        "summary": "Resolve Hold",
        "operationId": "resolve_hold_api_v1_confirmations_holds__hold_id__resolve_post",
        "parameters": [
          {
            "name": "hold_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Hold Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldResolve"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldResolveResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmationCreate": {
        "properties": {
          "client_ref": {
            "type": "string",
            "format": "uuid",
            "title": "Client Ref"
          },
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "shift": {
            "type": "string",
            "enum": [
              "A",
              "B"
            ],
            "title": "Shift"
          },
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "good_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Good Qty"
          },
          "rejected_qty": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Rejected Qty",
            "default": "0"
          },
          "reject_reason_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason Id"
          },
          "downtime_min": {
            "type": "integer",
            "title": "Downtime Min",
            "default": 0
          },
          "downtime_reason_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downtime Reason Id"
          },
          "process_loss": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessLoss"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "interim",
              "shift_close"
            ],
            "title": "Kind",
            "default": "interim"
          },
          "shift_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Date"
          }
        },
        "type": "object",
        "required": [
          "client_ref",
          "line_id",
          "shift",
          "material_id",
          "good_qty"
        ],
        "title": "ConfirmationCreate",
        "description": "POST /confirmations body \u2014 also the frozen payload of a confirmation_hold\n(\u00a711.11), so resolve_hold can re-validate it with this exact model."
      },
      "ConfirmationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "shift": {
            "type": "string",
            "title": "Shift"
          },
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "good_qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Good Qty"
          },
          "rejected_qty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Rejected Qty"
          },
          "reject_reason_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason Id"
          },
          "downtime_min": {
            "type": "integer",
            "title": "Downtime Min"
          },
          "downtime_reason_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downtime Reason Id"
          },
          "process_loss": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Loss"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "client_ref": {
            "type": "string",
            "format": "uuid",
            "title": "Client Ref"
          },
          "supervisor_id": {
            "type": "integer",
            "title": "Supervisor Id"
          },
          "posted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Posted At"
          },
          "shift_context_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Context Id"
          },
          "posted_after_close": {
            "type": "boolean",
            "title": "Posted After Close"
          },
          "sap_sync_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sap Sync Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "line_id",
          "shift",
          "material_id",
          "good_qty",
          "rejected_qty",
          "reject_reason_id",
          "downtime_min",
          "downtime_reason_id",
          "process_loss",
          "kind",
          "status",
          "client_ref",
          "supervisor_id",
          "posted_at",
          "shift_context_id",
          "posted_after_close"
        ],
        "title": "ConfirmationRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "material_id": {
            "type": "integer",
            "title": "Material Id"
          },
          "shift_context_id": {
            "type": "integer",
            "title": "Shift Context Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resolved_order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Order Id"
          },
          "client_ref": {
            "type": "string",
            "format": "uuid",
            "title": "Client Ref"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "line_id",
          "material_id",
          "shift_context_id",
          "payload",
          "status",
          "resolved_order_id",
          "client_ref",
          "created_by",
          "created_at"
        ],
        "title": "HoldRead"
      },
      "HoldResolve": {
        "properties": {
          "sap_order_no": {
            "type": "string",
            "title": "Sap Order No"
          },
          "line_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Id"
          },
          "material_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Id"
          }
        },
        "type": "object",
        "required": [
          "sap_order_no"
        ],
        "title": "HoldResolve",
        "description": "PPC supplies the SAP order; line/material default from the hold itself."
      },
      "HoldResolveResult": {
        "properties": {
          "hold": {
            "$ref": "#/components/schemas/HoldRead"
          },
          "confirmation": {
            "$ref": "#/components/schemas/ConfirmationRead"
          }
        },
        "type": "object",
        "required": [
          "hold",
          "confirmation"
        ],
        "title": "HoldResolveResult"
      },
...
      "ProcessLoss": {
        "properties": {
          "blanking": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Blanking",
            "default": "0"
          },
          "piercing": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Piercing",
            "default": "0"
          },
          "forming": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Forming",
            "default": "0"
          }
        },
        "type": "object",
        "title": "ProcessLoss",
        "description": "\u00a74.7 process_loss JSON \u2014 per-stage loss quantities."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__production__CorrectionCreate": {
        "properties": {
          "delta_good": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Delta Good"
          },
          "delta_reject": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Delta Reject",
            "default": "0"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "delta_good",
          "reason"
        ],
        "title": "CorrectionCreate"
      },
      "app__schemas__production__CorrectionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "confirmation_id": {
            "type": "integer",
            "title": "Confirmation Id"
          },
          "delta_good": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Delta Good"
          },
          "delta_reject": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Delta Reject"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "approval_id": {
            "type": "integer",
            "title": "Approval Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "confirmation_id",
          "delta_good",
          "delta_reject",
          "reason",
          "approval_id",
          "status"
        ],
        "title": "CorrectionRead"
      }
    }
  }
}